+++ src/FilterBundle/BestItCommercetoolsFilterBundle.php
@@ -2,6 +2,8 @@
 
 namespace BestIt\Commercetools\FilterBundle;
 
+use BestIt\Commercetools\FilterBundle\DependencyInjection\CompilerPass\TermNormalizerPass;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
 /**
@@ -13,4 +15,13 @@ use Symfony\Component\HttpKernel\Bundle\Bundle;
  */
 class BestItCommercetoolsFilterBundle extends Bundle
 {
+    /**
+     * {@inheritdoc}
+     */
+    public function build(ContainerBuilder $container)
+    {
+        parent::build($container);
+
+        $container->addCompilerPass(new TermNormalizerPass());
+    }
 }