+++ src/FilterBundle/Factory/FacetCollectionFactory.php
@@ -2,16 +2,18 @@
 
 namespace BestIt\Commercetools\FilterBundle\Factory;
 
-use BestIt\Commercetools\FilterBundle\Event\Facet\TermEvent;
-use BestIt\Commercetools\FilterBundle\FilterEvent;
+use BestIt\Commercetools\FilterBundle\Exception\SkipTermException;
 use BestIt\Commercetools\FilterBundle\Model\Facet\Facet;
 use BestIt\Commercetools\FilterBundle\Model\Facet\FacetCollection;
+use BestIt\Commercetools\FilterBundle\Model\Facet\FacetConfig;
 use BestIt\Commercetools\FilterBundle\Model\Facet\FacetConfigCollection;
 use BestIt\Commercetools\FilterBundle\Model\Facet\RangeCollection;
+use BestIt\Commercetools\FilterBundle\Model\Normalizer\TermNormalizerCollection;
+use BestIt\Commercetools\FilterBundle\Model\Search\SearchContext;
 use BestIt\Commercetools\FilterBundle\Model\Term\Term;
 use BestIt\Commercetools\FilterBundle\Model\Term\TermCollection;
+use BestIt\Commercetools\FilterBundle\Normalizer\TermNormalizerInterface;
 use Commercetools\Core\Model\Product\FacetResultCollection;
-use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 
 /**
  * Factory for facets
@@ -30,36 +32,41 @@ class FacetCollectionFactory
     private $facetConfigCollection;
 
     /**
-     * The event dispatcher
+     * The normalizer collection
      *
-     * @var EventDispatcherInterface
+     * @var TermNormalizerCollection
      */
-    private $eventDispatcher;
+    private $normalizerCollection;
 
     /**
      * FacetCollectionFactory constructor.
      *
      * @param FacetConfigCollection $facetConfigCollection
-     * @param EventDispatcherInterface $eventDispatcher
+     * @param TermNormalizerCollection $normalizerCollection
      */
-    public function __construct(FacetConfigCollection $facetConfigCollection, EventDispatcherInterface $eventDispatcher)
-    {
+    public function __construct(
+        FacetConfigCollection $facetConfigCollection,
+        TermNormalizerCollection $normalizerCollection
+    ) {
         $this->facetConfigCollection = $facetConfigCollection;
-        $this->eventDispatcher = $eventDispatcher;
+        $this->normalizerCollection = $normalizerCollection;
     }
 
     /**
      * Create facets collection
      *
      * @param FacetResultCollection $resultCollection
+     * @param SearchContext $searchContext
      *
      * @return FacetCollection
      */
-    public function create(FacetResultCollection $resultCollection): FacetCollection
+    public function create(FacetResultCollection $resultCollection, SearchContext $searchContext): FacetCollection
     {
         $result = $resultCollection->toArray();
 
         $collection = new FacetCollection();
+
+        /** @var FacetConfig $config */
         foreach ($this->facetConfigCollection as $config) {
             if (!isset($result[$config->getAlias()])) {
                 continue;
@@ -70,13 +77,20 @@ class FacetCollectionFactory
             $termsCollection = new TermCollection();
             if (isset($facet['terms'])) {
                 foreach ($facet['terms'] as $term) {
-                    $event = new TermEvent(
-                        $config,
-                        new Term($term['count'], $term['term'] ?? null, $term['term'] ?? null)
-                    );
+                    $termModel = new Term($term['count'], $term['term'] ?? null, $term['term'] ?? null);
 
-                    $this->eventDispatcher->dispatch(FilterEvent::FACET_TERM_COLLECT, $event);
-                    $termsCollection->addTerm($event->getTerm());
+                    try {
+                        /** @var TermNormalizerInterface $normalizer */
+                        foreach ($this->normalizerCollection as $normalizer) {
+                            if ($normalizer->support($config->getType())) {
+                                $termModel = $normalizer->normalize($config, $termModel, $searchContext);
+                            }
+                        }
+                    } catch (SkipTermException $exception) {
+                        continue;
+                    }
+
+                    $termsCollection->addTerm($termModel);
                 }
             }